==> data_preprocessing/ipynb/all_scripts/01_WPS.ipynb <==
## 라이브러리

import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
import matplotlib as mpl
import warnings
warnings.filterwarnings('ignore')

## path 설정

path = '../..'
script_path = f'{path}/script'
word_path = f'{path}/words/final_datasets'

## 결과 데이터프레임 생성

if os.path.exists(f'{script_path}/csv_files/all_scripts_WPS.csv'):
    df_result = pd.read_csv(f'{script_path}/csv_files/all_scripts_WPS.csv', index_col='title')
else:
    df_result = pd.DataFrame(columns=['title', 'WPS_mean', 'WPS_std', 'WPS_min', 'WPS_max']).set_index('title')

df_result


## 파일 불러오기

import os

# 다큐멘터리(0), 드라마(1), 영화(2) 선택
category = 1
if category == 0:
    script_path_cat = script_path + '/Documentary'
elif category == 1:
    script_path_cat = script_path + '/Drama'
else:
    script_path_cat = script_path + '/Movie'

_list = os.listdir(script_path_cat)
file_list = []
for file in _list:
    if file.startswith('.'): continue
    file_list.append(file)

file_list = sorted(file_list)
print(f'작품개수: {len(file_list)}')
count = 0
for index, title in enumerate(file_list):
    print(f'[{index}: {title}]', end=', ')
    count += 1
    if count%5 == 0:
        print()

## 여기서 작품 선택합니다 꼭!!

# 바로 위 코드에서 file_list를 참고하여 작품 선택
title = file_list[0]
if (df_result.index==title).any(): print(f'{title} is already exists')

file_name = title + '.WEBRip.Netflix.en[cc]'
print(f'{script_path_cat}/{title}/{file_name}')

df_file = pd.read_csv(f'{script_path_cat}/{title}/{file_name}.csv')
df = df_file
display(df.info(), df)

## 자막이 재생되는 시간 계산

df['start'] = pd.to_datetime(df['start'])
df['end'] = pd.to_datetime(df['end'])
df

df['duration'] = df['end'] - df['start']
df['duration'] = df['duration'].astype('str')
df['duration'] = df['duration'].str.split(' ').str[2]
df['duration'] = df['duration'].str.split(':').str[2].astype('float')
display(df.info(), df.head())

## 필요없는 컬럼 제거

df = df.drop(columns=['start', 'end'])
df.head()

## 단어의 개수 계산(띄어쓰기 기준)


df['words_num'] = df['script'].apply(lambda x: len(x.split(' ')))
df

# 1초에 몇 개의 단어가 보여지는지 계산

df['wordsPerSec'] = df['words_num'] / df['duration']
df['secPerWord'] = df['duration'] / df['words_num']
df

## 이상치 제거
### 1. 1개의 단어로 된 문장 제거
### 2. Words Per Sec 값이 0.5 이하인 문장 제거

df = df[df['words_num']!=1]
df = df[df['wordsPerSec']>=0.5].sort_values('words_num')
df

## describe 보기

df.describe()

df = df.sort_values('wordsPerSec')
pd.set_option('min_rows', 10)
pd.set_option('max_rows', 10)
display(df.head(10), df.tail(10))

df = df.sort_values('secPerWord')
df.tail(10)

result = df['wordsPerSec'].describe().to_frame().transpose()[['mean', 'std', 'min', 'max']]
result['title'] = title
result = result.set_index('title')
result = result.rename(columns={'mean':'WPS_mean', 'std':'WPS_std', 'min':'WPS_min', 'max':'WPS_max'})
result

## df_result에 합치기

df_result = pd.concat([df_result, result])
df_result

# 위의 코드들 합쳐서 한 번에 모든 작품 돌리기

import os
import pandas as pd
import numpy as np
# import seaborn as sns
# import matplotlib.pyplot as plt
# import matplotlib as mpl
# import warnings
# warnings.filterwarnings('ignore')

path = '../..'
# path = '/content/drive/MyDrive/elice_데이터분석_프로젝트/팀_공유_폴더'
script_path = f'{path}/script'
word_path = f'{path}/words/final_datasets'

if os.path.exists(f'{script_path}/csv_files/all_scripts_WPS.csv'):
    df_result = pd.read_csv(f'{script_path}/csv_files/all_scripts_WPS.csv', index_col='title')
else:
    df_result = pd.DataFrame(columns=['title', 'WPS_mean', 'WPS_std', 'WPS_min', 'WPS_max']).set_index('title')

display(df_result)

# 다큐멘터리(0), 드라마(1), 영화(2) 선택
category = 2
if category == 0:
    script_path_cat = script_path + '/Documentary'
elif category == 1:
    script_path_cat = script_path + '/Drama'
else:
    script_path_cat = script_path + '/Movie'

_list = os.listdir(script_path_cat)
file_list = []
for file in _list:
    if file.startswith('.'): continue
    file_list.append(file)

file_list = sorted(file_list)
print(f'작품개수: {len(file_list)}')
count = 0
for index, title in enumerate(file_list):
    print(f'[{index}: {title}]', end=', ')
    count += 1
    if count%5 == 0:
        print()


for i in range(len(file_list)):
    # 바로 위 코드에서 file_list를 참고하여 작품 선택
    title = file_list[i]
    if (df_result.index==title).any(): continue

    file_name = title + '.WEBRip.Netflix.en[cc]'
    # print(f'{path}/{title}/{file_name}')

    df_file = pd.read_csv(f'{script_path_cat}/{title}/{file_name}.csv')
    df = df_file

    df['start'] = pd.to_datetime(df['start'])
    df['end'] = pd.to_datetime(df['end'])

    df['duration'] = df['end'] - df['start']
    df['duration'] = df['duration'].astype('str')
    df['duration'] = df['duration'].str.split(' ').str[2]
    df['duration'] = df['duration'].str.split(':').str[2].astype('float')

    df = df.drop(columns=['start', 'end'])

    df['words_num'] = df['script'].apply(lambda x: len(x.split(' ')))

    df['wordsPerSec'] = df['words_num'] / df['duration']
    df['secPerWord'] = df['duration'] / df['words_num']

    df = df[df['words_num']!=1]
    df = df[df['wordsPerSec']>=0.5].sort_values('words_num')

    result = df['wordsPerSec'].describe().to_frame().transpose()[['mean', 'std', 'min', 'max']]
    result['title'] = title
    result = result.set_index('title')
    result = result.rename(columns={'mean':'WPS_mean', 'std':'WPS_std', 'min':'WPS_min', 'max':'WPS_max'})

    df_result = pd.concat([df_result, result])

df_result

## CSV 파일로 변환

df_result = df_result.sort_index()
df_result.to_csv(f'{script_path}/csv_files/all_scripts_WPS.csv')
