==> data_preprocessing/ipynb/all_scripts/02_running_time_for_vtt.ipynb <==
## 라이브러리

import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
import matplotlib as mpl
import re
import csv
import warnings
warnings.filterwarnings('ignore')

plt.rcParams['font.family'] = 'AppleGothic'
# plt.rcParams['font.family'] = 'NanumGothic'

## vvt 파일 불러오기

import os
path = '../../script'

# 다큐멘터리(0), 영화(1), 드라마(2) 선택
category = 0
if category == 0:
    path = path + '/Documentary'
    type = 'Documentary'
elif category == 1:
    path = path + '/Movie'
    type = 'Movie'
elif category == 2:
    path = path + '/Drama'
    type = 'Drama'   

_list = os.listdir(path)
file_list = []
for file in _list:
    if file.startswith('.'): continue
    file_list.append(file)

file_list = sorted(file_list)
print(f'작품개수: {len(file_list)}')
# print([{i:_list[i]} for i in range(len(file_list))])
for index, title in enumerate(file_list):
    print(f'{index}: {title}')


def get_running_time_sec(title, type):

    path = f'../script/{type}'
    file_name = title + '.WEBRip.Netflix.en[cc]'
    if type == "Drama":
        title = title.split('.S0')[0]
    print(f'{path}/{title}/{file_name}')

    # vtt 파일 읽기
    opened_file = open(f'{path}/{title}/{file_name}.vtt', encoding='utf8')
    content = opened_file.read()
    segments = content.split('\n\n')    # 한줄씩 정리

    # 불필요한 부분 제거
    m = re.compile(r"\<.*?\>")
    def clean(content):
        if '♪' in content: return None
        if '[' in content: return None
        if ']' in content: return None
        if '(' in content: return None
        if ')' in content: return None
        if '{' in content: return None
        if '}' in content: return None
        new_content = m.sub('',content)
        new_content = new_content.replace('-->','')
        remove_point_1 = new_content.find('position:')
        remove_point_2 = new_content.find('% \n')
        new_content = new_content[:remove_point_1] + new_content[remove_point_2+3:]
        return new_content

    new_segments = [clean(s) for s in segments if len(s)!=0][4:]

    # 줄번호 / 속도 시작 / 속도 끝 / 자막 - 데이터 분리
    trimmed_segments = []
    for segment in new_segments:
        if segment != None and segment.strip() != '':
            split_segment = segment.split()
            times = split_segment[:3]
            script = ' '.join(split_segment[3:])
            times.append(script)
            trimmed_segments.append(times)

    df = pd.DataFrame(columns=['num', 'start', 'end', 'script'])

    for row in trimmed_segments:
        df.loc[row[0]] = row

    df['start'] = pd.to_datetime(df['start'])
    df['end'] = pd.to_datetime(df['end'])

    duration = df['end'][len(df)-1] - df['start'][0]

    return duration.total_seconds()



table = pd.read_csv('../../script/csv_files/all_scripts_running_time.csv')

if type == 'Documentary' or type == 'Movie':
    for title in file_list:
        row = []
        row.append(title)
        row.append(get_running_time_sec(title, type))
        row.append(3600/row[1])
        table.loc[title] = row

if type == 'Drama':
    for title in file_list:
        drama_list = os.listdir('../../script/Drama/' + title)
        episode_list = []
        for episode in drama_list:
            if '.vtt' in episode:
                episode = episode.split('.WEBRip.')[0]
                episode_list.append(episode)
        
        totol_running_time = 0
        for episode in episode_list:
            totol_running_time += get_running_time_sec(episode, type)
        
        row = []
        row.append(title)
        row.append(totol_running_time)
        row.append(3600/row[1])
        table.loc[title] = row

table

# table.to_csv('../../script/csv_files/all_scripts_running_time.csv', index=False)
table.to_csv('../../script/csv_files/all_scripts_running_time.csv', index=False)

hi = pd.read_csv('../../script/csv_files/all_scripts_running_time.csv')
hi

_list = os.listdir('../../script/Drama')
file_list = []
for file in _list:
    if file.startswith('.'): continue
    file_list.append(file)

drama_list = os.listdir('../../script/Drama/' + file_list[0])
episode_list = []
for episode in _list:
    if '.vtt' in episode:
        episode = episode.split('.WEBRip.')[0]
        episode_list.append(episode)

episode_list
