==> data_preprocessing/ipynb/all_scripts/03_merge.ipynb <==
## 라이브러리

import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
import matplotlib as mpl
import warnings
warnings.filterwarnings('ignore')

## path 설정

path = '../..'
# path = 
script_path = f'{path}/script/csv_files'

## 데이터 프레임 불러오기

df_result_word_level_counts = pd.read_csv(f'{script_path}/all_scripts_word_level_counts.csv', index_col='title')
df_result_WPS = pd.read_csv(f'{script_path}/all_scripts_WPS.csv', index_col='title')
df_result_running_time = pd.read_csv(f'{script_path}/all_scripts_running_time.csv', index_col='title')


## 데이터 확인

display(df_result_word_level_counts.info(), df_result_word_level_counts)

display(df_result_WPS.info(), df_result_WPS)

display(df_result_running_time.info(), df_result_running_time)

df_result = df_result_word_level_counts.merge(df_result_WPS, on='title')
df_result = df_result.merge(df_result_running_time, on='title')
df_result

## csv 파이롤 변환

df_result = df_result.sort_index()
df_result.to_csv(f'{script_path}/all_scripts.csv')
